+++ lib/distribute_reads/global_methods.rb
@@ -10,7 +10,9 @@ module DistributeReads
       previous_value = Thread.current[:distribute_reads]
       begin
         Thread.current[:distribute_reads] = true
-        yield
+        value = yield
+        warn "[distribute_reads] Call `to_a` inside block to execute query on replica" if value.is_a?(ActiveRecord::Relation) && !previous_value
+        value
       ensure
         Thread.current[:distribute_reads] = previous_value
       end